+++ src/private-messages/controllers/queries/private-messages.move-to-folder.query.ts
@@ -1,0 +1,7 @@
+import { IsEnum } from 'class-validator';
+import { PrivateMessageFolder } from 'src/private-messages/types';
+
+export class PrivateMessagesMoveToFolderQuery {
+  @IsEnum(PrivateMessageFolder)
+  folder: PrivateMessageFolder;
+}