+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [2.1.0] - unrelease
+
+### Added
+
+- Implemented `/privateMessages/:id/markAsUnread` endpoint.
+- Implemented `/privateMessages/:id/moveToFolder` endpoint.
+
+### Fixed
+
+- Added missing error documentation to `/privateMessages/` endpoints.
+
 ## [2.0.2] - 2023-10-12
 
 ### Fixed